+++ controllers/website.js
@@ -1,18 +1,15 @@
 const db = require("../models"); // models path depend on your structure
+const response = require('../lib/response');
 
 // create website detail
 const create = async (req, res, next) => {
     try {
         const websiteData = req.body;
         await db.website.create(websiteData)
-        res.send({
-            message: 'website create successfully'
-        })
-    } catch(e) {
-        res.status(500).send({
-            message:
-                e.message || "Some error occurred while create website."
-        });
+        res.send(response.success('Website has been created successfully',{}))
+        
+    } catch(err) {
+        res.status(err.status || 422).send(response.error(err.errors));       
     }
 }
 
@@ -20,15 +17,15 @@ const create = async (req, res, next) => {
 const update = async (req, res, next) => {
     try {
         const websiteData = req.body;
+
+        if(!websiteData.id){
+            throw {status:422, errors:{message:'Id is required'}}
+        }
         await db.website.update(websiteData, { where: { id: websiteData.id } })
-        res.send({
-            message: 'website update successfully'
-        })
-    } catch(e) {
-        res.status(500).send({
-            message:
-                e.message || "Some error occurred while update website."
-        });
+       
+        res.send(response.success('Website has been updated successfully',{}))
+    } catch(err) {
+        res.status(err.status || 422).send(response.error(err.errors)); 
     }
 }
 
@@ -36,28 +33,34 @@ const update = async (req, res, next) => {
 const getAll = async (req, res, next) => {
     try {
         const websites = await db.website.findAll()
-        res.send(websites)
-    } catch (e) {
-        res.status(500).send({
-            message:
-                e.message || "Some error occurred while get websites."
-        });
+    
+        res.send(response.success('Websites listing',websites))
+    } catch (err) {
+        res.status(err.status || 422).send(response.error(err.errors)); 
     }
 }
 
 // delete website details
 const deletes = async (req, res, next) => {
     try {
-        const id = req.query.id;
+
+        if(!req.body.id){
+            throw {status:422, errors:{message:'Id is required'}}
+        }
+
+        const id = req.body.id;
         const website = await db.website.destroy({where: {id: id}})
-        res.send({
-            message: 'delete website successfully'
-        })
-    } catch(e) {
-        res.status(500).send({
-            message:
-                e.message || "Some error occurred while delete website."
-        });
+        console.log('website', website)
+
+        if(website){
+            res.send(response.success('Website has been deleted successfully',{}))
+        } else {
+            throw {status:422, errors:{message:'Website is not found'}}
+        }
+     
+        
+    } catch(err) {
+        res.status(err.status || 422).send(response.error(err.errors)); 
     }
     
 }
